In-memory write caching (#7)
* fix: comment typo

* fix: add eslint-config-standard-kit for linting/formatting

* Add eslint-config-standard-kit, eslint, and prettier
* Add lint-staged
* Upgrade to Typescript 4.2.4
* Add `yarn lint` and `yarn fix` scripts

* fix: code reorg

* fix: list method implementation

Instead of a direct pass-through to Disklet, Memlet needs its own
implementation because to be able to do in-memory write caching.

* fix: rename filename to key for File and FileMap types

* fix: rename FileQueue queue method to enqueue

* fix: no dash file name convention

* fix: memory leak fix for lastTouchedTimestamp

A file's timestamp should not be modified before requeue. Doing so will
break binary search in the file queue.

* fix: replace FileQueue with generic Queue

* fix: remove lastTouchedTimestamp; internalize this state in the Queue

To avoid the bugs that can arise from mutating lastTouchedTimestamp,
we should instead leave the timestamp as an implementation detail in
the queue.

* fix(tests): add 'will not leak memory in queues' test

* fix: move delay util to src/

* feat: in-memory write caching

* fixup! feat: in-memory write caching

Add drainTimeoutId to prevent more than one instance of the routine running

* fixup! feat: in-memory write caching

Remove blank lines and cleanup with comments

* fixup! feat: in-memory write caching

Using an actionQueue

* fix(queue): binary search on key equality

Use the QueueItem key as the identifier for object equality in the
queue. This makes it simpler to reason about.

* fixup! feat: in-memory write caching

May exceed memory limit until action queue is flushed

* fixup! feat: in-memory write caching

Use a generator for flush event

* fixup! feat: in-memory write caching

This comment is not true anymore. Also, no need to even expose _nextFlushEvent now that you have onFlush

* fixup! feat: in-memory write caching

no reason to use the underscore convention since you don't even expose this param.

* fixup! feat: in-memory write caching

Can you please "one liner" it? ;)

* fixup! feat: in-memory write caching

Unfortunately we can't use Map in any of the plugins or the the libs that are being used them.
It has some unpredictable bugs after being transpiled in the App.