Rewrite Action type to fix missing delete action bug
Delete actions would not be queued when a file is in the memory cache.
To fix this we must always queue a delete action even if we don't have
a reference to the file object.
We must refactor the Action type to hold any generic action routine so
we can prepare an action with or without a File object reference.
This simplifies our API and unblocks us from writing the fix.